Implement HIV in encryption/decryption.
Nonce/ctr state is no longer necessary. Simply use a small hash of the
message as the intial value of the old nonce and use counter starting
from zero, which guarantees that a nonce+ctr pair will never be used
twice unless its the same message, in which case no extra information is
revealed besides that some same message was sent twice.